requirements.txt
numpy
pandas
matplotlib
seaborn
python-dotenv~=1.0.1
graphviz
requests
PyMySQL~=1.1.0